+++ src/m365/aad/commands/app/app-add.mock.ts
@@ -4928,3 +4928,49 @@ export const aadSp = {
   "oauth2PermissionScopes": [],
   "appRoles": []
 };
+export const mockCrmSp = {
+  
+  "servicePrincipalNames": [
+    "https://admin.services.crm17.dynamics.com/",
+    "00000007-0000-0000-c000-000000000000/*.crm9.dynamics.com",
+    "00000007-0000-0000-c000-000000000000/*.crm7.dynamics.com",
+    "00000007-0000-0000-c000-000000000000/*.crm6.dynamics.com",
+    "00000007-0000-0000-c000-000000000000/*.crm2.dynamics.com",
+    "00000007-0000-0000-c000-000000000000/*.crm5.dynamics.com",
+    "00000007-0000-0000-c000-000000000000/*.crm4.dynamics.com",
+    "00000007-0000-0000-c000-000000000000/*.crm.dynamics.com",
+    "00000007-0000-0000-c000-000000000000",
+    "https://admin.services.crm.dynamics.com/",
+    "https://admin.services.crm2.dynamics.com/",
+    "https://admin.services.crm3.dynamics.com/",
+    "https://admin.services.crm4.dynamics.com/",
+    "https://admin.services.crm5.dynamics.com/",
+    "https://admin.services.crm6.dynamics.com/",
+    "https://admin.services.crm7.dynamics.com/",
+    "https://admin.services.crm8.dynamics.com/",
+    "https://admin.services.crm9.dynamics.com/",
+    "https://admin.services.crm11.dynamics.com/",
+    "https://admin.services.crm.microsoftdynamics.us/",
+    "https://admin.services.crm.appsplatform.us/",
+    "https://admin.services.crm12.dynamics.com/",
+    "https://admin.services.crm14.dynamics.com/",
+    "https://admin.services.crm15.dynamics.com/",
+    "https://admin.services.crm16.dynamics.com/",
+    "https://siteapi.crm2.crmlivetie.com/"
+],
+"appId": "00000007-0000-0000-c000-000000000000",
+"oauth2PermissionScopes": [
+    {
+        "adminConsentDescription": "Allows the application to access Common Data Service acting as users in the organization.",
+        "adminConsentDisplayName": "Access Common Data Service as organization users",
+        "id": "78ce3f0f-a1ce-49c2-8cde-64b5c0896db4",
+        "isEnabled": true,
+        "type": "User",
+        "userConsentDescription": "Allows the application to access Common Data Service as you.",
+        "userConsentDisplayName": "Access Common Data Service as you",
+        "value": "user_impersonation"
+    }
+],
+"appRoles": []
+};
+