+++ src/m365/aad/commands/app/app-add.spec.ts
@@ -19,6 +19,13 @@ describe(commands.APP_ADD, () => {
     sinon.stub(auth, 'restoreAuth').callsFake(() => Promise.resolve());
     sinon.stub(appInsights, 'trackEvent').callsFake(() => { });
     auth.service.connected = true;
+    auth.service.tenantId = '48526e9f-60c5-3000-31d7-aa1dc75ecf3c|908bel80-a04a-4422-b4a0-883d9847d110:c8e761e2-d528-34d1-8776-dc51157d619a&#xA;Tenant';
+    if (!auth.service.accessTokens[auth.defaultResource]) {
+      auth.service.accessTokens[auth.defaultResource] = {
+        expiresOn: 'abc',
+        value: 'abc'
+      };
+    }
   });
 
   beforeEach(() => {
@@ -148,7 +155,8 @@ describe(commands.APP_ADD, () => {
         assert.strictEqual(typeof err, 'undefined');
         assert(loggerLogSpy.calledWith({
           appId: 'bc724b77-da87-43a9-b385-6ebaaf969db8',
-          objectId: '5b31c38c-2584-42f0-aa47-657fb3a84230'
+          objectId: '5b31c38c-2584-42f0-aa47-657fb3a84230',
+          tenantId: ''
         }));
         done();
       }
@@ -246,7 +254,8 @@ describe(commands.APP_ADD, () => {
         assert.strictEqual(typeof err, 'undefined');
         assert(loggerLogSpy.calledWith({
           appId: '62f0f128-987f-47f2-827a-be50d0d894c7',
-          objectId: '9b1e2c08-6e35-4134-a0ac-16ab154cd05a'
+          objectId: '9b1e2c08-6e35-4134-a0ac-16ab154cd05a',
+          tenantId: ''
         }));
         done();
       }
@@ -354,7 +363,8 @@ describe(commands.APP_ADD, () => {
         assert.strictEqual(typeof err, 'undefined');
         assert(loggerLogSpy.calledWith({
           appId: 'd2941a3b-aad4-49e0-8a1d-b82de0b46067',
-          objectId: 'ff520671-4810-4d25-a10f-e565fc62a5ec'
+          objectId: 'ff520671-4810-4d25-a10f-e565fc62a5ec',
+          tenantId: ''
         }));
         done();
       }
@@ -460,7 +470,8 @@ describe(commands.APP_ADD, () => {
         assert.strictEqual(typeof err, 'undefined');
         assert(loggerLogSpy.calledWith({
           appId: '1ce0287c-9ccc-457e-a0cf-3ec5b734c092',
-          objectId: 'f1bb2138-bff1-491e-b082-9f447f3742b8'
+          objectId: 'f1bb2138-bff1-491e-b082-9f447f3742b8',
+          tenantId: ''
         }));
         done();
       }
@@ -572,7 +583,8 @@ describe(commands.APP_ADD, () => {
         assert(loggerLogSpy.calledWith({
           appId: '3c5bd51d-f1ac-4344-bd16-43396cadff14',
           objectId: '4d24b0c6-ad07-47c6-9bd8-9c167f9f758e',
-          secret: 'VtJt.yG~V5pzbY2.xekx_0Xy_~9ozP_Ub5'
+          secret: 'VtJt.yG~V5pzbY2.xekx_0Xy_~9ozP_Ub5',
+          tenantId: ''
         }));
         done();
       }
@@ -684,7 +696,8 @@ describe(commands.APP_ADD, () => {
         assert(loggerLogSpy.calledWith({
           appId: '3c5bd51d-f1ac-4344-bd16-43396cadff14',
           objectId: '4d24b0c6-ad07-47c6-9bd8-9c167f9f758e',
-          secret: 'VtJt.yG~V5pzbY2.xekx_0Xy_~9ozP_Ub5'
+          secret: 'VtJt.yG~V5pzbY2.xekx_0Xy_~9ozP_Ub5',
+          tenantId: ''
         }));
         done();
       }
@@ -843,7 +856,8 @@ describe(commands.APP_ADD, () => {
         assert(loggerLogSpy.calledWith({
           appId: 'dbfdad7a-5105-45fc-8290-eb0b0b24ac58',
           objectId: 'b63c4be1-9c78-40b7-8619-de7172eed8de',
-          secret: 'vP2K-_K-N6EI-E5z0yOTsz443grfM_pyvv'
+          secret: 'vP2K-_K-N6EI-E5z0yOTsz443grfM_pyvv',
+          tenantId: ''
         }));
         done();
       }
@@ -1011,7 +1025,8 @@ describe(commands.APP_ADD, () => {
         assert(loggerLogSpy.calledWith({
           appId: 'dbfdad7a-5105-45fc-8290-eb0b0b24ac58',
           objectId: 'b63c4be1-9c78-40b7-8619-de7172eed8de',
-          secret: 'vP2K-_K-N6EI-E5z0yOTsz443grfM_pyvv'
+          secret: 'vP2K-_K-N6EI-E5z0yOTsz443grfM_pyvv',
+          tenantId: ''
         }));
         done();
       }
@@ -1173,7 +1188,8 @@ describe(commands.APP_ADD, () => {
         assert.strictEqual(typeof err, 'undefined');
         assert(loggerLogSpy.calledWith({
           appId: 'c505d465-9e4e-4bb4-b653-7b36d77cc94a',
-          objectId: 'f51ff52f-8f04-4924-91d0-636349eed65c'
+          objectId: 'f51ff52f-8f04-4924-91d0-636349eed65c',
+          tenantId: ''
         }));
         done();
       }
@@ -1335,7 +1351,8 @@ describe(commands.APP_ADD, () => {
         assert.strictEqual(typeof err, 'undefined');
         assert(loggerLogSpy.calledWith({
           appId: 'c505d465-9e4e-4bb4-b653-7b36d77cc94a',
-          objectId: 'f51ff52f-8f04-4924-91d0-636349eed65c'
+          objectId: 'f51ff52f-8f04-4924-91d0-636349eed65c',
+          tenantId: ''
         }));
         done();
       }
@@ -1444,7 +1461,8 @@ describe(commands.APP_ADD, () => {
         assert.strictEqual(typeof err, 'undefined');
         assert(loggerLogSpy.calledWith({
           appId: 'b08d9318-5612-4f87-9f94-7414ef6f0c8a',
-          objectId: 'c0e63919-057c-4e6b-be6c-8662e7aec4eb'
+          objectId: 'c0e63919-057c-4e6b-be6c-8662e7aec4eb',
+          tenantId: ''
         }));
         done();
       }
@@ -1553,7 +1571,8 @@ describe(commands.APP_ADD, () => {
         assert.strictEqual(typeof err, 'undefined');
         assert(loggerLogSpy.calledWith({
           appId: 'b08d9318-5612-4f87-9f94-7414ef6f0c8a',
-          objectId: 'c0e63919-057c-4e6b-be6c-8662e7aec4eb'
+          objectId: 'c0e63919-057c-4e6b-be6c-8662e7aec4eb',
+          tenantId: ''
         }));
         done();
       }
@@ -1680,7 +1699,8 @@ describe(commands.APP_ADD, () => {
         assert.strictEqual(typeof err, 'undefined', `Error: ${JSON.stringify(err)}`);
         assert(loggerLogSpy.calledWith({
           appId: '13e11551-2967-4985-8c55-cd2aaa6b80ad',
-          objectId: 'fe45ba27-a692-4b11-adf8-f4ec184ea3a5'
+          objectId: 'fe45ba27-a692-4b11-adf8-f4ec184ea3a5',
+          tenantId: ''
         }));
         done();
       }
@@ -1807,7 +1827,8 @@ describe(commands.APP_ADD, () => {
         assert.strictEqual(typeof err, 'undefined', `Error: ${JSON.stringify(err)}`);
         assert(loggerLogSpy.calledWith({
           appId: '13e11551-2967-4985-8c55-cd2aaa6b80ad',
-          objectId: 'fe45ba27-a692-4b11-adf8-f4ec184ea3a5'
+          objectId: 'fe45ba27-a692-4b11-adf8-f4ec184ea3a5',
+          tenantId: ''
         }));
         done();
       }
@@ -2375,6 +2396,153 @@ describe(commands.APP_ADD, () => {
     }, (err?: any) => {
       try {
         assert.strictEqual(JSON.stringify(err), JSON.stringify(new CommandError('An error has occurred')));
+        done();
+      }
+      catch (e) {
+        done(e);
+      }
+    });
+  });
+
+  it('creates AAD app reg for a web app with service principal name with trailing slash', (done) => {
+    sinon.stub(request, 'get').callsFake(opts => {
+      if (opts.url === 'https://graph.microsoft.com/v1.0/myorganization/servicePrincipals?$select=servicePrincipalNames,appId,oauth2PermissionScopes,appRoles') {
+        return Promise.resolve({
+          "@odata.nextLink": "https://graph.microsoft.com/v1.0/myorganization/servicePrincipals?$select=servicePrincipalNames%2cappId%2coauth2PermissionScopes%2cappRoles&$skiptoken=X%274453707402000100000035536572766963655072696E636970616C5F34623131646566352D626561622D343232382D383835622D61323963386536336638613235536572766963655072696E636970616C5F34623131646566352D626561622D343232382D383835622D6132396338653633663861320000000000000000000000%27",
+          "value": [
+            mocks.mockCrmSp
+          ]
+        });
+      }
+
+      if (opts.url === 'https://graph.microsoft.com/v1.0/myorganization/servicePrincipals?$select=servicePrincipalNames%2cappId%2coauth2PermissionScopes%2cappRoles&$skiptoken=X%274453707402000100000035536572766963655072696E636970616C5F34623131646566352D626561622D343232382D383835622D61323963386536336638613235536572766963655072696E636970616C5F34623131646566352D626561622D343232382D383835622D6132396338653633663861320000000000000000000000%27') {
+        return Promise.resolve({
+          value: mocks.aadSp
+        });
+      }
+
+      return Promise.reject(`Invalid GET request: ${opts.url}`);
+    });
+    sinon.stub(request, 'patch').callsFake(_ => Promise.reject('Issued PATCH request'));
+    sinon.stub(request, 'post').callsFake(opts => {
+      if (opts.url === 'https://graph.microsoft.com/v1.0/myorganization/applications' &&
+        JSON.stringify(opts.data) === JSON.stringify({
+          "displayName": "My AAD app",
+          "signInAudience": "AzureADMyOrg",
+          "requiredResourceAccess": [
+            {
+              "resourceAppId": "00000007-0000-0000-c000-000000000000",
+              "resourceAccess": [
+                {
+                  "id": "78ce3f0f-a1ce-49c2-8cde-64b5c0896db4",
+                  "type": "Scope"
+                }
+              ]
+            }
+          ],
+          "web": {
+            "redirectUris": [
+              "https://global.consent.azure-apim.net/redirect"
+            ]
+          }
+        })) {
+        return Promise.resolve({
+          "id": "1cd23c5f-2cb4-4bd0-a582-d5b00f578dcd",
+          "deletedDateTime": null,
+          "appId": "702e65ba-cacb-4a2f-aa5c-e6460967bc20",
+          "applicationTemplateId": null,
+          "createdDateTime": "2021-02-21T09:44:05.953701Z",
+          "displayName": "My AAD app",
+          "description": null,
+          "groupMembershipClaims": null,
+          "identifierUris": [],
+          "isDeviceOnlyAuthSupported": null,
+          "isFallbackPublicClient": null,
+          "notes": null,
+          "optionalClaims": null,
+          "publisherDomain": "m365404404.onmicrosoft.com",
+          "signInAudience": "AzureADMyOrg",
+          "tags": [],
+          "tokenEncryptionKeyId": null,
+          "verifiedPublisher": {
+            "displayName": null,
+            "verifiedPublisherId": null,
+            "addedDateTime": null
+          },
+          "defaultRedirectUri": null,
+          "addIns": [],
+          "api": {
+            "acceptMappedClaims": null,
+            "knownClientApplications": [],
+            "requestedAccessTokenVersion": null,
+            "oauth2PermissionScopes": [],
+            "preAuthorizedApplications": []
+          },
+          "appRoles": [],
+          "info": {
+            "logoUrl": null,
+            "marketingUrl": null,
+            "privacyStatementUrl": null,
+            "supportUrl": null,
+            "termsOfServiceUrl": null
+          },
+          "keyCredentials": [],
+          "parentalControlSettings": {
+            "countriesBlockedForMinors": [],
+            "legalAgeGroupRule": "Allow"
+          },
+          "passwordCredentials": [],
+          "publicClient": {
+            "redirectUris": []
+          },
+          "requiredResourceAccess": [
+            {
+              "resourceAppId": "00000007-0000-0000-c000-000000000000",
+              "resourceAccess": [
+                {
+                  "id": "78ce3f0f-a1ce-49c2-8cde-64b5c0896db4",
+                  "type": "Scope"
+                }
+              ]
+            }
+          ],
+          "web": {
+            "homePageUrl": null,
+            "logoutUrl": null,
+            "redirectUris": [
+              "https://global.consent.azure-apim.net/redirect"
+            ],
+            "implicitGrantSettings": {
+              "enableAccessTokenIssuance": false,
+              "enableIdTokenIssuance": false
+            }
+          },
+          "spa": {
+            "redirectUris": []
+          }
+
+        });
+      }
+
+      return Promise.reject(`Invalid POST request: ${JSON.stringify(opts, null, 2)}`);
+    });
+
+    command.action(logger, {
+      options: {
+        debug: false,
+        name: 'My AAD app',
+        platform: 'web',
+        redirectUris: 'https://global.consent.azure-apim.net/redirect',
+        apisDelegated: 'https://admin.services.crm.dynamics.com/user_impersonation'
+      }
+    }, (err?: any) => {
+      try {
+        assert.strictEqual(typeof err, 'undefined');
+        assert(loggerLogSpy.calledWith({
+          appId: '702e65ba-cacb-4a2f-aa5c-e6460967bc20',
+          objectId: '1cd23c5f-2cb4-4bd0-a582-d5b00f578dcd',
+          tenantId: ''
+        }));
         done();
       }
       catch (e) {